Split benchmark for special operators
